+++ pjlib-util/src/pjlib-util/base64.c
@@ -119,7 +119,7 @@ PJ_DEF(pj_status_t) pj_base64_encode(const pj_uint8_t *input, int in_len,
 
 
 PJ_DEF(pj_status_t) pj_base64_decode(const pj_str_t *input, 
-				     pj_uint8_t *out, long *out_len)
+				     pj_uint8_t *out, int *out_len)
 {
     const char *buf = input->ptr;
     int len = input->slen;
